+++ CHANGES.rst
@@ -12,6 +12,7 @@ in progress
 - Improve software tests
 - Unlock reading from stdin
 - Derive table name from input file name
+- Primary key detection
 
 
 2023-01-xx 0.0.0